Reorganize library components; simplify and rename main user-facing interface
