+++ src/main/scala/vectorpipe/internal/package.scala
@@ -16,7 +16,7 @@ import org.locationtech.geomesa.spark.jts._
 import vectorpipe.functions.osm._
 import vectorpipe.relations.{MultiPolygons, Routes}
 
-object ProcessOSM {
+package object internal {
   val NodeType: Byte = 1
   val WayType: Byte = 2
   val RelationType: Byte = 3
@@ -58,25 +58,6 @@ object ProcessOSM {
       Nil)
 
   lazy val VersionedElementEncoder: Encoder[Row] = RowEncoder(VersionedElementSchema)
-
-  /**
-    * Snapshot pre-processed elements.
-    *
-    * A Time Pin is stuck through a set of elements that have been augmented with a 'validUntil column to identify all
-    * that were valid at a specific point in time (i.e. updated before the target timestamp and valid after it).
-    *
-    * @param df        Elements (including 'validUntil column)
-    * @param timestamp Optional timestamp to snapshot at
-    * @return DataFrame containing valid elements at timestamp (or now)
-    */
-  def snapshot(df: DataFrame, timestamp: Timestamp = null): DataFrame = {
-    import df.sparkSession.implicits._
-
-    df
-      .where(
-        'updated <= coalesce(lit(timestamp), current_timestamp)
-          and coalesce(lit(timestamp), current_timestamp) < coalesce('validUntil, date_add(current_timestamp, 1)))
-  }
 
   /**
     * Pre-process nodes.
@@ -215,35 +196,6 @@ object ProcessOSM {
           'version,
           'visible)
     }
-  }
-
-  /**
-    * Construct all geometries for a set of elements.
-    *
-    * This currently produces Points for nodes containing "interesting" tags, LineStrings and Polygons for ways
-    * (according to OSM rules for defining areas), MultiPolygons for multipolygon and boundary relations, and
-    * LineStrings / MultiLineStrings for route relations.
-    *
-    * @param elements DataFrame containing node, way, and relation elements
-    * @return DataFrame containing geometries.
-    */
-  def constructGeometries(elements: DataFrame): DataFrame = {
-    import elements.sparkSession.implicits._
-    val st_pointToGeom = org.apache.spark.sql.functions.udf { pt: jts.Point => pt.asInstanceOf[jts.Geometry] }
-
-    val nodes = ProcessOSM.preprocessNodes(elements)
-
-    val nodeGeoms = ProcessOSM.constructPointGeometries(nodes)
-      .withColumn("minorVersion", lit(0))
-      .withColumn("geom", st_pointToGeom('geom))
-
-    val wayGeoms = ProcessOSM.reconstructWayGeometries(elements, nodes)
-
-    val relationGeoms = ProcessOSM.reconstructRelationGeometries(elements, wayGeoms)
-
-    nodeGeoms
-      .union(wayGeoms.where(size('tags) > 0).drop('geometryChanged))
-      .union(relationGeoms)
   }
 
   /**
@@ -651,23 +603,6 @@ object ProcessOSM {
         'visible,
         'version,
         'minorVersion)
-  }
-
-  /**
-    * Augment geometries with user metadata.
-    *
-    * When 'changeset is included, user (name and 'uid) metadata is joined from a DataFrame containing changeset
-    * metadata.
-    *
-    * @param geoms      Geometries to augment.
-    * @param changesets Changesets DataFrame with user metadata.
-    * @return Geometries augmented with user metadata.
-    */
-  def addUserMetadata(geoms: DataFrame, changesets: DataFrame): DataFrame = {
-    import geoms.sparkSession.implicits._
-
-    geoms
-      .join(changesets.select('id as 'changeset, 'uid, 'user), Seq("changeset"))
   }
 
 }
